Checkpoint commit of CUDA graph progress (Friday).

+++ src/PerfectHash/GraphImpl1.h
@@ -176,13 +176,13 @@ RegisterEdgeDeletion(
     _In_ EDGE Edge
     )
 {
-    ULONG OrderIndex;
+    LONG OrderIndex;
     ASSERT(!TestGraphBit(DeletedEdgesBitmap, Edge));
     SetGraphBit(DeletedEdgesBitmap, Edge);
     Graph->DeletedEdgeCount++;
     ASSERT(Graph->DeletedEdgeCount <= Graph->TotalNumberOfEdges);
     OrderIndex = --Graph->OrderIndex;
-    ASSERT((LONG)OrderIndex >= 0);
+    ASSERT(OrderIndex >= 0);
     Graph->Order[OrderIndex] = Edge;
 }
